requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
peft>=0.6.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
audioread>=3.0.0

# Whisper specialized
openai-whisper>=20231117
faster-whisper>=0.9.0

# Evaluation
jiwer>=3.0.0
sentencepiece>=0.1.99

# Optimization
onnxruntime>=1.16.0
optimum>=1.14.0
bitsandbytes>=0.41.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.66.0
pyyaml>=6.0
tensorboard>=2.14.0
wandb>=0.15.0

# Commercial transcription services (optional)
# Install only what you need:
# assemblyai>=0.28.0  # For AssemblyAI
# deepgram-sdk>=3.2.0  # For Deepgram
# azure-cognitiveservices-speech>=1.32.0  # For Azure
# google-cloud-speech>=2.20.0  # For Google

# Development
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0

